adding forget and reset pass forms
