+++ frontend/src/components/ResetPasswordForm.js
@@ -1,0 +1,52 @@
+import React, { useState } from 'react';
+import { useParams, useNavigate } from 'react-router-dom';
+import axios from 'axios';
+
+const ResetPasswordForm = () => {
+  const { token } = useParams();
+  const navigate = useNavigate();
+  const [password, setPassword] = useState('');
+  const [confirm, setConfirm] = useState('');
+  const [error, setError] = useState('');
+  const [message, setMessage] = useState('');
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    if (password !== confirm) return setError("Passwords don't match");
+
+    try {
+      const res = await axios.post(`${process.env.REACT_APP_BACKEND_URL || 'https://dishcraft-backend-3tk2.onrender.com'}/api/reset-password/${token}`, { password });
+      setMessage(res.data.message);
+      setTimeout(() => navigate('/login'), 2500);
+    } catch (err) {
+      setError(err.response?.data?.message || 'Reset failed');
+    }
+  };
+
+  return (
+    <div>
+      <h2>Reset Password</h2>
+      <form onSubmit={handleSubmit}>
+        <input
+          type="password"
+          placeholder="New password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          required
+        />
+        <input
+          type="password"
+          placeholder="Confirm password"
+          value={confirm}
+          onChange={(e) => setConfirm(e.target.value)}
+          required
+        />
+        <button type="submit">Reset Password</button>
+      </form>
+      {message && <p style={{ color: 'green' }}>{message}</p>}
+      {error && <p style={{ color: 'red' }}>{error}</p>}
+    </div>
+  );
+};
+
+export default ResetPasswordForm;
